requirements.txt
flask==3.0.3
anthropic==0.40.0
playwright==1.48.0
requests==2.32.3
pdfplumber==0.11.4
Werkzeug==3.0.6
python-dotenv==1.0.1
apscheduler==3.10.4
